FlexControlNet/requirements.txt
torch
transformers
pillow
tqdm
torchvision
numpy
opencv-python
diffusers 
accelerate 
transformers 
bitsandbytes 
einops 
xformers==0.0.25 
nvidia-cutlass 
huggingface_hub 
datasets 
timm==1.0.9 
controlnet-aux==0.0.7 
torchmetrics  
torch-fidelity